Updated 2/5 10:22 pm
Can you fix the resume so it is centered and the same size as the image below. Also can you change the review for the screening questions because I removed some of the questions in the code. And what are we doing about the pop ups for the cancel and submit?